Rename the packages from io.quarkus.ai to io.quarkus.ai.harness
