--- tests/src/main/java/io/quarkus/ai/checks/CheckConfig.java
@@ -1,11 +1,0 @@
-package io.quarkus.ai.checks;
-
-import java.util.List;
-
-public record CheckConfig(
-        List<EndpointCheck> endpoints
-) {
-    public CheckConfig {
-        if (endpoints == null) endpoints = List.of();
-    }
-}
