--- tests/src/main/java/io/quarkus/ai/checks/EndpointCheck.java
@@ -1,19 +1,0 @@
-package io.quarkus.ai.checks;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-public record EndpointCheck(
-        String path,
-        String method,
-        String body,
-        @JsonProperty("expected_status") Integer expectedStatus,
-        @JsonProperty("body_contains") String bodyContains
-) {
-    public String effectiveMethod() {
-        return method == null || method.isBlank() ? "GET" : method.toUpperCase();
-    }
-
-    public int effectiveExpectedStatus() {
-        return expectedStatus != null ? expectedStatus : 200;
-    }
-}
