+++ tests/src/main/java/io/quarkus/ai/harness/checks/CheckConfig.java
@@ -1,0 +1,11 @@
+package io.quarkus.ai.harness.checks;
+
+import java.util.List;
+
+public record CheckConfig(
+        List<EndpointCheck> endpoints
+) {
+    public CheckConfig {
+        if (endpoints == null) endpoints = List.of();
+    }
+}
